requirements.txt
streamlit==1.0.0
pandas==0.25.1
Pillow==8.4.0
